requirements.txt
# Core Django Framework
Django~=6.0.1

# Django Extensions & Utilities
django-crispy-forms~=2.1
crispy-bootstrap4~=2025.6
crispy-bootstrap5~=2025.6
django-bootstrap4~=26.1
django-tables2~=2.8.0
django-select2~=8.4.8
django-filter~=25.2
django-jsoneditor~=0.2.4
django-appconf~=1.2.0

# Database
psycopg2==2.9.11

# Task Queue / Background Jobs
celery~=5.6.2
redis~=7.1.1

# AI / LLM Providers
generic-llm-api-client==0.3.3

# Data Processing & Analysis
pandas~=3.0.0
fuzzywuzzy==0.18.0
python-Levenshtein~=0.27.3

# HTTP & API Clients
requests~=2.32.0
google-api-python-client==2.190.0

# XML/HTML Parsing
lxml~=6.0.2
xmltodict~=1.0.2
beautifulsoup4~=4.14.3
simple-alto-parser==0.0.22

# Text Processing
Markdown~=3.10.2
python-decouple==3.8
dateparser~=1.3.0

# Geolocation
geopy~=2.4.1

# Semantic Web / RDF
SPARQLWrapper~=2.0.0
rdflib~=7.6.0

# Documentation (Optional - for building docs)
Sphinx~=9.1.0
docutils~=0.22.4

# Utilities
python-dateutil~=2.9.0.post0
pytz~=2025.2
PyYAML~=6.0.2
edtf~=5.0.0
